Move psbench.csv to psbench.results
