+++ psbench/benchmarks/endpoint_qps/main.py
@@ -20,9 +20,9 @@ from proxystore.connectors.endpoint import EndpointConnector
 
 from psbench.argparse import add_logging_options
 from psbench.benchmarks.endpoint_qps import routes
-from psbench.csv import CSVLogger
 from psbench.logging import init_logging
 from psbench.logging import TESTING_LOG_LEVEL
+from psbench.results import CSVLogger
 
 PROCESS_STARTUP_BUFFER_SECONDS = 5
 ROUTE_TYPE = Literal['GET', 'SET', 'EXISTS', 'EVICT', 'ENDPOINT']
