+++ psbench/benchmarks/remote_ops/main.py
@@ -26,9 +26,9 @@ from proxystore.p2p.relay.client import RelayClient
 import psbench.benchmarks.remote_ops.endpoint_ops as endpoint_ops
 import psbench.benchmarks.remote_ops.redis_ops as redis_ops
 from psbench.argparse import add_logging_options
-from psbench.csv import CSVLogger
 from psbench.logging import init_logging
 from psbench.logging import TESTING_LOG_LEVEL
+from psbench.results import CSVLogger
 
 BACKEND_TYPE = Literal['ENDPOINT', 'REDIS']
 OP_TYPE = Literal['EVICT', 'EXISTS', 'GET', 'SET']
