+++ psbench/benchmarks/workflow/main.py
@@ -23,13 +23,13 @@ from proxystore.store.scopes import submit
 from psbench.argparse import add_executor_options
 from psbench.argparse import add_logging_options
 from psbench.argparse import add_proxystore_options
-from psbench.csv import CSVLogger
 from psbench.executor.factory import init_executor_from_args
 from psbench.executor.protocol import Executor
 from psbench.logging import init_logging
 from psbench.logging import TESTING_LOG_LEVEL
 from psbench.memory import SystemMemoryProfiler
 from psbench.proxystore import init_store_from_args
+from psbench.results import CSVLogger
 from psbench.utils import randbytes
 
 logger = logging.getLogger('workflow')
